% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedercplus-package.R
\docType{package}
\name{pedercplus-package}
\alias{pedercplus}
\alias{pedercplus-package}
\title{pedercplus: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pedersebastian/pedercplus}
  \item \url{https://pedersebastian.github.io/pedercplus/}
  \item Report bugs at \url{https://github.com/pedersebastian/pedercplus/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
