% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_cpp}
\alias{sum_cpp}
\title{Summer alle elementer i en numerisk vektor}
\usage{
sum_cpp(x)
}
\arguments{
\item{x}{En numerisk vektor}
}
\value{
Summen av alle elementer
}
\description{
Summer alle elementer i en numerisk vektor
}
