% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_lower_v1}
\alias{to_lower_v1}
\title{Versjon 1: std::transform med ::tolower}
\usage{
to_lower_v1(x)
}
\arguments{
\item{x}{En tekststreng}
}
\value{
Strengen med små bokstaver
}
\description{
Versjon 1: std::transform med ::tolower
}
