% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_lower_v3}
\alias{to_lower_v3}
\title{Versjon 3: range-based for-loop med referanse}
\usage{
to_lower_v3(x)
}
\arguments{
\item{x}{En tekststreng}
}
\value{
Strengen med små bokstaver
}
\description{
Versjon 3: range-based for-loop med referanse
}
