% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar1_cpp}
\alias{ar1_cpp}
\title{Simuler en AR(1)-prosess}
\usage{
ar1_cpp(n, phi, sigma)
}
\arguments{
\item{n}{Antall tidssteg}

\item{phi}{Autokorrelasjon (-1 til 1). 0.9 = treg mean-reversion}

\item{sigma}{Standardavvik på støyen}
}
\value{
Numerisk vektor av lengde n
}
\description{
Hvert steg: x(t) = phi * x(t-1) + støy.
Fordi hvert steg avhenger av forrige kan ikke R vektorisere dette.
}
\examples{
set.seed(1)
x <- ar1_cpp(100, phi = 0.9, sigma = 1)
plot(x, type = "l")
}
