% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedercplus-package.R
\docType{package}
\name{pedercplus-package}
\alias{pedercplus}
\alias{pedercplus-package}
\title{pedercplus: Learning C++ and Rcpp Through Examples}
\description{
A personal learning package exploring C++ and Rcpp. Contains implementations of string manipulation and time series simulation, used to compare different C++ approaches and understand when C++ offers performance benefits over plain R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pedersebastian/pedercplus}
  \item \url{https://pedersebastian.github.io/pedercplus/}
  \item Report bugs at \url{https://github.com/pedersebastian/pedercplus/issues}
}

}
\author{
\strong{Maintainer}: Peder Martinsen \email{peder.s.martinsen@gmail.com}

}
\keyword{internal}
