% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_lower_v2}
\alias{to_lower_v2}
\title{Versjon 2: manuell loop med ASCII-aritmetikk
Store bokstaver A-Z har ASCII 65-90, små a-z har 97-122 (avstand 32)}
\usage{
to_lower_v2(x)
}
\arguments{
\item{x}{En tekststreng}
}
\value{
Strengen med små bokstaver
}
\description{
Versjon 2: manuell loop med ASCII-aritmetikk
Store bokstaver A-Z har ASCII 65-90, små a-z har 97-122 (avstand 32)
}
